Fix Kbar in operator mode
