+++ portal-ui/src/screens/Console/valid-routes.ts
@@ -32,6 +32,7 @@ import {
   GroupsMenuIcon,
   HealthMenuIcon,
   IdentityMenuIcon,
+  InspectMenuIcon,
   LogsMenuIcon,
   MetricsMenuIcon,
   MonitoringMenuIcon,
@@ -39,7 +40,6 @@ import {
   SupportMenuIcon,
   TraceMenuIcon,
   UsersMenuIcon,
-  InspectMenuIcon,
 } from "../../icons/SidebarMenus";
 import { hasPermission } from "../../common/SecureComponent";
 import WatchIcon from "../../icons/WatchIcon";
@@ -59,7 +59,6 @@ export const validRoutes = (
   operatorMode: boolean
 ) => {
   const ldapIsEnabled = (features && features.includes("ldap-idp")) || false;
-
   let consoleMenus: IMenuItem[] = [
     {
       name: "Buckets",
@@ -159,13 +158,6 @@ export const validRoutes = (
           icon: DrivesMenuIcon,
           component: NavLink,
         },
-        {
-          name: "Inspect",
-          id: "inspectObjects",
-          to: IAM_PAGES.TOOLS_INSPECT,
-          icon: InspectMenuIcon,
-          component: NavLink,
-        },
       ],
     },
     {
@@ -202,13 +194,13 @@ export const validRoutes = (
         //   icon: CallHomeMenuIcon,
         //   to: IAM_PAGES.CALL_HOME,
         // },
-        // {
-        //   name: "Inspect",
-        //   id: "inspect",
-        //   component: NavLink,
-        //   icon: InspectMenuIcon,
-        //   to: IAM_PAGES.TOOLS_WATCH,
-        // },
+        {
+          name: "Inspect",
+          id: "inspectObjects",
+          to: IAM_PAGES.TOOLS_INSPECT,
+          icon: InspectMenuIcon,
+          component: NavLink,
+        },
         // {
         //   name: "Profile",
         //   id: "profile",
@@ -277,6 +269,7 @@ export const validRoutes = (
     {
       group: "Operator",
       type: "item",
+      id: "Tenants",
       component: NavLink,
       to: IAM_PAGES.TENANTS,
       name: "Tenants",
@@ -286,6 +279,7 @@ export const validRoutes = (
     {
       group: "Operator",
       type: "item",
+      id: "License",
       component: NavLink,
       to: IAM_PAGES.LICENSE,
       name: "License",
@@ -295,6 +289,7 @@ export const validRoutes = (
     {
       group: "Operator",
       type: "item",
+      id: "Documentation",
       component: NavLink,
       to: IAM_PAGES.DOCUMENTATION,
       name: "Documentation",
